notegraph: fix warnings.

+++ game/notegraph.hh
@@ -26,8 +26,8 @@ class NoteGraph {
 	void drawNotes();
 	/// draw waves (what players are singing)
 	void drawWaves(Database const& database);
- 	double barHeight();
- 	double waveThickness();
+ 	float barHeight();
+ 	float waveThickness();
 	VocalTrack const& m_vocal;
 	Texture m_notelines;
 	Texture m_wave;
@@ -43,7 +43,7 @@ class NoteGraph {
 	AnimValue m_nlTop, m_nlBottom;
 	Notes::const_iterator m_songit;
 	double m_time;
-	double m_max, m_min, m_noteUnit, m_baseY, m_baseX;
+	float m_max, m_min, m_noteUnit, m_baseY, m_baseX;
 	const NoteGraphScalerPtr m_scaler;
 };
